chore : #49 - utils index 로 이동
