+++ src/hooks/useCurrentPage.ts
@@ -1,7 +1,7 @@
 /* eslint-disable react-hooks/rules-of-hooks */
 import { useLocation, useMatch } from 'react-router-dom';
 import { routes } from '@routes';
-import { parseQueryString } from '@utils/parseQueryString';
+import { parseQueryString } from '@utils';
 
 interface SearchParams {
   show?: 'true';
