--- src/utils/parseQueryString.ts
@@ -1,9 +1,0 @@
-export const parseQueryString = (queryString: string) => {
-	return queryString
-		.slice(1)
-		.split('&')
-		.reduce((obj, query) => {
-			const [key, value] = query.split('=');
-			return { ...obj, [key]: value };
-		}, {});
-};
